+++ output/outelf32.c
@@ -772,7 +772,7 @@ static int32_t elf_add_gsym_reloc(struct Section *sect,
     return offset - sym->value;
 }
 
-static void elf_out(int32_t segto, const void *data, uint32_t type,
+static void elf_out(int32_t segto, const void *data, uint64_t type,
                     int32_t segment, int32_t wrt)
 {
     struct Section *s;
